big changes; units are now SI from vasp chgcar units, entry points has had a lot of little tweaks -- no need to pass filename to bader-read fortran-format is used by export also & completely new jit caching, to facilitate the new caching a new file called jits.py has been created with dictionaries of sample input for each jitted function in each module, only the volumes array is dynamically typed as it is the larges memory hog and typing the otheres makes little difference to mem usage but large difference to required storage.

+++ pybader/methods.py
@@ -56,9 +56,9 @@ def ongrid(density, volumes, idx, dist_mat, T_grad, i_c):
     path_num = 0
     # init arrays for bader maxima, edge crossings and current path
     # idx is type set for the maximum int that can be stored here
-    bader_max = np.zeros((vx, 3), dtype=idx.dtype)
-    edge_max = np.zeros((vx, 3), dtype=idx.dtype)
-    path = np.zeros((vx, 3), dtype=idx.dtype)
+    bader_max = np.zeros((vx, 3), dtype=np.int64)
+    edge_max = np.zeros((vx, 3), dtype=np.int64)
+    path = np.zeros((vx, 3), dtype=np.int64)
     # init position arrays
     p = np.zeros(3, dtype=np.int64)
     pt = np.zeros(3, dtype=np.int64)
@@ -262,9 +262,9 @@ def neargrid(density, volumes, idx, dist_mat, T_grad, i_c):
     path_num = 0
     # init arrays for bader maxima, edge crossings and current path
     # idx is type set for the maximum int that can be stored here
-    bader_max = np.zeros((vx, 3), dtype=idx.dtype)
-    edge_max = np.zeros((vx, 3), dtype=idx.dtype)
-    path = np.zeros((vx, 3), dtype=idx.dtype)
+    bader_max = np.zeros((vx, 3), dtype=np.int64)
+    edge_max = np.zeros((vx, 3), dtype=np.int64)
+    path = np.zeros((vx, 3), dtype=np.int64)
     # init position arrays
     p = np.zeros(3, dtype=np.int64)
     pd = np.zeros(3, dtype=np.int64)
